nav-item

+++ src/main/java/com/kobe/warehouse/service/dto/records/VenteModePaimentRecord.java
@@ -1,5 +1,4 @@
 package com.kobe.warehouse.service.dto.records;
 
-import java.math.BigDecimal;
 
-public record VenteModePaimentRecord(String code, String libelle, BigDecimal netAmount, BigDecimal paidAmount) {}
+public record VenteModePaimentRecord(String code, String libelle, Integer netAmount, Integer paidAmount) {}
